pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Muhammad Iqbal <iqbal.adudu@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
django = "^4.1.5"
djangorestframework = "^3.14.0"
pillow = "^9.4.0"
django-rest-registration = "^0.7.3"
coverage = "^7.0.5"
django-cors-headers = "^3.13.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
